has_silent_letters has been dividied into several methods and methods to determine same sound letters
docstring was updated for all the methods of silent letters. The code in
has_silent_letters was divided into has_silent_h and has_silent_u so
it's easier to maintain.

the algorithm to determine same sound letters was added; It comprises 7
methods and their docstring has been added (no code yet).