Some of the code for checking anagrams has been added
The basic code for check anagram has been added. If an instance of
IOData doesn't exist, it creates one, so it can read the anagrams data
base.

to do:
Not sure about the nested for loop (need to check if one of them can be
run in another function.

a method that cleans the line from the anagram data base needs to be
added. It should split the str into a list, no its easier to manipulate.